LoginForm
Отделение LoginForm от App
